Fixed BIG soundness problem.
We assumed that the qualifier of an abstraction fun['a] -> e should be
the same as the qualifier of the body of the abstraction, e.  Thus, we
define |all 'a. t| = |t|.  But type abstractions are closures, so we
need to make sure that they don't improperly close over affine
variables.  We weren't checking this at all, so we admitted erroneous
programs such as:

  let[A] dummy =
    let r      = ref 4 in
    let f['a]  = let (_, z) = swap (r, true) in z in
    let (a, b) = (f[unit], f[unit]) in
      a + b

This program attempts to add an integer to a boolean.

Two potential fixes:

 1. Type abstractions, like abstractions, need to be annotated with a
 qualifier:  |all^q 'a. t| = q.  Then, we ensure that the qualifier
 given to any type abstraction is no weaker than the qualifiers of any
 of its closed-over variables (just like with lambdas).  This is the
 solution adoptied by Fluet's rgnURAL.  This change affects all programs
 that contain polymorphic types, since their syntax changes.

 2. Allow only values in the body of type abstractions, not arbitrarily
 effectful expressions.  Then, we can safely assume that the qualifier of
 an abstraction equals the qualifier of its body, since copying a
 type-abstracted value is just as sound as copying that value.  In
 particular, in terms of the form "fun['a] (x: t) -> e", the free
 variable check that determines the qualifier of the inner lambda is
 inherited by the outer type lambda.  This doesn't diminish
 expressiveness, since we can always add an additional lambda
 abstraction under any type-lambda without changing the meaning of the
 program.

I chose 2, since it offers the chance that programs need fewer changes
overall, and I'd rather keep the types simpler.  In fact, only 3
programs (some with variations) needed changes.  In one case, it was
sufficient to eta-expand a partial application:

  fun['a] -> f x   ===>   fun['a] (lst: 'a list) -> f x lst

In the remaining cases, it sufficed to re-order the arguments of a
function to move some final type-arguments before a final
value-argument.

This highlights an interesting difference between our approach and
Matthew's, in which his may (I'm not sure) be more expressive, whereas
ours comes out much simpler.  In the *URAL systems, each type
and value constructor more or less carries a qualifier with it.  For
example, to type products, the rule is sort of like:

   G |- ei : ti    |t1| <= q    |t2| <= q
   --------------------------------------
   G |- q^(e1, e2) : q^(t1 * t2)

That is, a pair gets its own qualifier, which must be at least as
restrictive as the qualifiers of its components.  We, on the other hand,
give pairs types as qualifier the least upper-bound of the qualifiers of
their two components, which is a tight bound.  Then, if we find that we
want to use a pair at an even more restrictive qualifier, subtyping
handles that.  This seems to eliminate the need for qualifier variables,
pre-types, constrained types, etc., for now.  I think we don't have
principal types, but *almost*, and adding a simple form of constraint
such as |t1| = |t2| may be sufficient.