When checking a mu type, makes sure that the type is well-formed in terms of type constructor bounds. For example, given the following definitions:
type 'a t [rec 'a]
type u = mu 'a. 'a t
type v = mu 'a. ('a * int aref) t
type w = mu 'a. 'a t * int aref

Both v and w should be rejected, because both amount to applying t to an
affine type (note that the sigil on mu-bound variables doesn't mean anything),
but it wasn't rejecting the w case, because the int aref doesn't appear
directly under t.  Now it sets up a constraint that rejects it *eventually*,
but I need to make it check satisfiability right away.